\name{ar}
\alias{ar}
\docType{data}
\title{ar data}
\usage{data(ar)}
\description{
The \code{ar} data frame has 60 rows and 4 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{x1}{a numeric vector}
    \item{x2}{a numeric vector}
    \item{x3}{a numeric vector}
    \item{y}{a numeric vector}
  }
}
%\details{}
%\source{}
\references{Atkinson, A.C. and Riani, M. (2000),  \emph{Robust Diagnostic Regression Analysis}, First Edition. New York: Springer, Table A.2}
%\examples{}
\keyword{datasets}
