\name{bliss}
\alias{bliss}
\docType{data}
\title{Bliss data}
\usage{data(bliss)}
\description{
The \code{bliss} data frame has 8 rows and 4 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Dose}{a numeric vector}
    \item{Killed}{a numeric vector}
    \item{Total}{a numeric vector}
    \item{y}{a numeric vector}
  }
}
\details{}
\source{}
\references{Atkinson, A.C. and Riani, M. (2000),  \emph{Robust Diagnostic Regression Analysis}, First Edition. New York: Springer, Table A.20}
\examples{}
\keyword{datasets}
