\name{calcium}
\alias{calcium}
\docType{data}
\title{Calcium data}
\usage{data(calcium)}
\description{
Calcium uptake of cells suspended in a solution of radioactive calcium. \cr
The \code{calcium} data frame has 27 rows and 2 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Time}{a numeric vector}
    \item{y}{a numeric vector}
  }
}
%\details{}
%\source{}
\references{Atkinson, A.C. and Riani, M. (2000),  \emph{Robust Diagnostic Regression Analysis}, First Edition. New York: Springer, Table A.13}
%\examples{}
\keyword{datasets}
