\name{carinsuk}
\alias{carinsuk}
\non_function{}
\title{Car insurance data}
\usage{data(carinsuk)}
\description{
The \code{carinsuk} data frame has 128 rows and 5 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{OwnerAge}{a factor with levels}
    \item{OwnerAge}{\code{17-20} }
    \item{OwnerAge}{\code{21-24} }
    \item{OwnerAge}{\code{25-29} }
    \item{OwnerAge}{\code{30-34} }
    \item{OwnerAge}{\code{35-39} }
    \item{OwnerAge}{\code{40-49} }
    \item{OwnerAge}{\code{50-59} }
    \item{OwnerAge}{\code{60+} }
    \item{Model}{a factor with levels}
    \item{Model}{\code{A} }
    \item{Model}{\code{B} }
    \item{Model}{\code{C} }
    \item{Model}{\code{D} }
    \item{CarAge}{a factor with levels}
    \item{CarAge}{\code{0-3} }
    \item{CarAge}{\code{10+} }
    \item{CarAge}{\code{4-7} }
    \item{CarAge}{\code{8-9} }
    \item{NClaims}{a numeric vector}
    \item{AvCost}{a numeric vector}
  }
}
\details{}
\source{}
\references{Atkinson, A.C. and Riani, M. (2000),  \emph{Robust Diagnostic Regression Analysis}, First Edition. New York: Springer, Table A.16}
\examples{}
\keyword{datasets}
