\name{carr}
\alias{carr}
\docType{data}
\title{n-Pentane data}
\usage{data(carr)}
\description{Reaction rate for Catalytic Isomerization of n-Pentane to
Isopentane \cr
The \code{carr} data frame has 24 rows and 4 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{hydrogen}{partial pressure of hydrogen}
    \item{npentane}{partial pressure of n-pentane}
    \item{isopentane}{partial pressure of iso-pentane}
    \item{rate}{rate of disappearance of n-pentane}
  }
}
%\details{}
%\source{}
\references{Atkinson, A.C. and Riani, M. (2000),  \emph{Robust Diagnostic Regression Analysis}, First Edition. New York: Springer, Table A.15}
%\examples{}
\keyword{datasets}
