\name{cellular}
\alias{cellular}
\non_function{}
\title{Cellular differentiation data}
\usage{data(cellular)}
\description{
The \code{cellular} data frame has 16 rows and 3 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{TNF}{Dose of TNF (U/ml) }
    \item{IFN}{Dose of IFN (U/ml) }
    \item{y}{Number of cells differentiating}
  }
}
\details{}
\source{}
\references{Atkinson, A.C. and Riani, M. (2000),  \emph{Robust Diagnostic Regression Analysis}, First Edition. New York: Springer, Table A.19}
\examples{}
\keyword{datasets}
