\name{chapman}
\alias{chapman}
\docType{data}
\title{Chapman data}
\usage{data(chapman)}
\description{
The \code{chapman} data frame has 200 rows and 7 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{age}{a numeric vector}
    \item{highbp}{a numeric vector}
    \item{lowbp}{a numeric vector}
    \item{chol}{a numeric vector}
    \item{height}{a numeric vector}
    \item{weight}{a numeric vector}
    \item{y}{a numeric vector}
  }
}
%\details{}
%\source{}
\references{Atkinson, A.C. and Riani, M. (2000),  \emph{Robust Diagnostic Regression Analysis}, First Edition. New York: Springer, Table A.24}
%\examples{}
\keyword{datasets}
