\name{dialectric}
\alias{dialectric}
\docType{data}
\title{Dialectric data}
\usage{data(dialectric)}
\description{
The \code{dialectric} data frame has 128 rows and 3 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{time}{Time (weeks)}
    \item{temp}{Temperature (degrees Celsius)}
    \item{y}{dialectric breakdown strength in kilovolts}
  }
}
\details{}
\source{}
\references{Atkinson, A.C. and Riani, M. (2000), \emph{Robust Diagnostic Regression Analysis}, First Edition. New York: Springer, Table A.17}
\examples{}
\keyword{datasets}
