% Generated by roxygen2 (4.0.2): do not edit by hand
\name{evolveStd}
\alias{evolveStd}
\title{Evolve a single population under multiplicative fitness and arbitrary changes in N}
\usage{
evolveStd(rng, N0, mu_n, mu_s, s, h, littler, Ns, dist = TRUE)
}
\arguments{
\item{rng}{A wrapper to a GSL random number generator}

\item{N0}{The initial population size}

\item{mu_n}{The neutral mutation rate (per gamete per generation)}

\item{s}{The selection coefficient (for fixed-s models), or the mean of an exponential (for models with a distribution)}

\item{h}{The dominance of a selected mutation}

\item{littler}{The recombination rate (per diploid, per generation).}

\item{Ns}{A list of populaton sizes.  The length of the list is the number of generations to simulate}

\item{dist}{If true, simulate a distribution of fitness effects, otherwise treat s as fixed}

\item{mu_n}{The mutation rate to selected mutations (per gamete per generation)}
}
\description{
Evolve a single population under multiplicative fitness and arbitrary changes in N
}

