% Generated by roxygen2 (4.0.2): do not edit by hand
\name{evolveWFsingle}
\alias{evolveWFsingle}
\title{Evolve a single population under multiplicative fitness and arbitrary changes in N}
\usage{
evolveWFsingle(pop, mu_n, mu_s, s, h, littler, Ns, seed, dist = TRUE)
}
\description{
Evolve a single population under multiplicative fitness and arbitrary changes in N
}

