\name{forbes}
\alias{forbes}
\non_function{}
\title{Forbes data}
\usage{data(forbes)}
\description{
Forbes' data on air pressure in the Alps and the boiling point of
water. \cr
The \code{forbes} data frame has 17 rows and 2 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{x}{Boiling point}
    \item{y}{100 × log(pressure)}
  }
}
\details{}
\source{}
\references{Atkinson, A.C. and Riani, M. (2000),  \emph{Robust Diagnostic Regression Analysis}, First Edition. New York: Springer, Table A.1}
\examples{}
\keyword{datasets}
