\name{Forward-package}
\alias{Forward-package}
\alias{Forward}
\docType{package}
\title{
\packageTitle{Forward}
}
\description{
\packageDescription{Forward}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Forward}
\packageIndices{Forward}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{Forward}

Maintainer: \packageMaintainer{Forward}
}
\keyword{ package }
