% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{foRward}
\alias{foRward}
\alias{foRward-package}
\title{foRward.}
\description{
An Rcpp/C++11 package for forward simulation using fwdpp
}
\author{
Kevin R. Thornton \email{krthornt@uci.edu}
}
\references{
\url{https://github.com/molpopgen/foRward}
}

