\name{hawkins}
\alias{hawkins}
\docType{data}
\title{Hawkins' data}
\usage{data(hawkins)}
\description{
The \code{hawkins} data frame has 128 rows and 9 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{x1}{a numeric vector}
    \item{x2}{a numeric vector}
    \item{x3}{a numeric vector}
    \item{x4}{a numeric vector}
    \item{x5}{a numeric vector}
    \item{x6}{a numeric vector}
    \item{x7}{a numeric vector}
    \item{x8}{a numeric vector}
    \item{y}{a numeric vector}
  }
}
\details{}
\source{}
\references{Atkinson, A.C. and Riani, M. (2000),  \emph{Robust Diagnostic Regression Analysis}, First Edition. New York: Springer, Table A.4}
\examples{}
\keyword{datasets}
