\name{kinetics}
\alias{kinetics}
\non_function{}
\title{Kinetics data}
\usage{data(kinetics)}
\description{Kinetics data (from Becton-Dickenson)  \cr
The \code{kinetics} data frame has 19 rows and 5 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Substrate}{substrate indicator}
    \item{I0}{Inhibitor concentration}
    \item{I3}{Inhibitor concentration}
    \item{I10}{Inhibitor concentration}    
    \item{I30}{Inhibitor concentration}
    \item{y}{initial velocity}
  }
}
\details{}
\source{}
\references{Atkinson, A.C. and Riani, M. (2000),  \emph{Robust Diagnostic Regression Analysis}, First Edition. New York: Springer, Table A.12}
\examples{}
\keyword{datasets}
