\name{lakes}
\alias{lakes}
\docType{data}
\title{Lakes data}
\usage{data(lakes)}
\description{
The \code{lakes} data frame has 29 rows and 3 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{NIN}{average influent nitrogenon concentration}
    \item{TW}{water retention time}
    \item{TN}{mean annual nitrogen concentration}
  }
}
%\details{}
%\source{}
\references{Atkinson, A.C. and Riani, M. (2000),  \emph{Robust Diagnostic Regression Analysis}, First Edition. New York: Springer, Table A.14}
%\examples{}
\keyword{datasets}
