\name{leafpine}
\alias{leafpine}
\docType{data}
\title{Pine data}
\usage{data(leafpine)}
\description{
The \code{leafpine} data frame has 70 rows and 3 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{girth}{girth}
    \item{height}{height}
    \item{volume}{volume}
  }
}
%\details{}
%\source{}
\references{Atkinson, A.C. and Riani, M. (2000),  \emph{Robust Diagnostic Regression Analysis}, First Edition. New York: Springer, Table A.10}
%\examples{}
\keyword{datasets}
