\name{mice}
\alias{mice}
\docType{data}
\title{Mice data}
\usage{data(mice)}
\description{
The \code{mice} data frame has 14 rows and 4 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{dose}{dose level}
    \item{prep}{factor preparation: 0= Standard preparation, 1= Test preparation}
    \item{conv}{number with convultion}
    \item{total}{Total}
  }
}
\details{}
\source{}
\references{Atkinson, A.C. and Riani, M. (2000),  \emph{Robust
Diagnostic Regression Analysis}, First Edition. New York: Springer,
Table A.21}
\examples{}
\keyword{datasets}
