\name{molar}
\alias{molar}
\docType{data}
\title{Molar data}
\usage{data(molar)}
\description{Radioactivity versus molar concentration of nifedipene \cr
The \code{molar} data frame has 15 rows and 2 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{x}{log10(NIF concentration)}
    \item{y}{Total counts for \eqn{5 \times 10^-10} Molar NTD additive}
  }
}
%\details{}
%\source{}
\references{Atkinson, A.C. and Riani, M. (2000),  \emph{Robust Diagnostic Regression Analysis}, First Edition. New York: Springer, Table A.1}
%\examples{}
\keyword{datasets}
