\name{mussels}
\alias{mussels}
\docType{data}
\title{Mussels data}
\usage{data(mussels)}
\description{
The \code{mussels} data frame has 82 rows and 5 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{W}{width}
    \item{H}{height}
    \item{L}{length}
    \item{S}{shell mass}
    \item{M}{mass}
  }
}
\details{}
\source{}
\references{Atkinson, A.C. and Riani, M. (2000),  \emph{Robust Diagnostic Regression Analysis}, First Edition. New York: Springer, Table A.9}
\examples{}
\keyword{datasets}
