\name{ozone}
\alias{ozone}
\docType{data}
\title{Ozone data}
\usage{data(ozone)}
\description{Ozone concentration at Upland, CA. \cr
The \code{ozone} data frame has 80 rows and 9 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{x1}{a numeric vector}
    \item{x2}{a numeric vector}
    \item{x3}{a numeric vector}
    \item{x4}{a numeric vector}
    \item{x5}{a numeric vector}
    \item{x6}{a numeric vector}
    \item{x7}{a numeric vector}
    \item{x8}{a numeric vector}
    \item{y}{ Ozone concentration (ppm)}
  }
}
\details{}
\source{}
\references{Atkinson, A.C. and Riani, M. (2000),  \emph{Robust Diagnostic Regression Analysis}, First Edition. New York: Springer, Table A.7}
\examples{}
\keyword{datasets}
