\name{poison}
\alias{poison}
\docType{data}
\title{Poison data}
\usage{data(poison)}
\description{Box and Cox poison data. Survival times in 10 hour units
of animals in a \eqn{3 \times 4}  factorial experiment. \cr
The \code{poison} data frame has 48 rows and 3 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{time}{a numeric vector}
    \item{poison}{a factor}
    \item{treat}{a factor with levels: \code{A}, \code{B}, \code{C}, \code{D} }
  }
}
%\details{}
%\source{}
\references{Atkinson, A.C. and Riani, M. (2000), \emph{Robust Diagnostic Regression Analysis}, First Edition. New York: Springer, Table A.8}
%\examples{}
\keyword{datasets}
