\name{rainfall}
\alias{rainfall}
\non_function{}
\title{Rainfall data}
\usage{data(rainfall)}
\description{Toxoplasmosis data. \cr
The \code{rainfall} data frame has 34 rows and 3 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Rain}{mm of rain}
    \item{Cases}{cases of toxoplasmosis}
    \item{Total}{total}
  }
}
\details{}
\source{}
\references{Atkinson, A.C. and Riani, M. (2000),  \emph{Robust Diagnostic Regression Analysis}, First Edition. New York: Springer, Table A.22}
\examples{}
\keyword{datasets}
