% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sample_std}
\alias{sample_std}
\title{Sample nsam chromosomes from population}
\usage{
sample_std(pop, nsam, seed)
}
\arguments{
\item{pop}{A population returned from evolve_std}

\item{nsam}{The sample size}

\item{seed}{RNG seed}
}
\description{
Sample nsam chromosomes from population
}

