% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sampleSingle}
\alias{sampleSingle}
\title{Sample nsam chromosomes from population}
\usage{
sampleSingle(pop, nsam, seed)
}
\arguments{
\item{pop}{A population returned from evolveWFsingle}

\item{nsam}{The sample size}

\item{seed}{RNG seed}
}
\description{
Sample nsam chromosomes from population
}

