% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sampleStd}
\alias{sampleStd}
\title{Sample nsam chromosomes from population}
\usage{
sampleStd(pop, rng, nsam)
}
\arguments{
\item{pop}{A population returned from evolve_std}

\item{rng}{A GSLrng}

\item{nsam}{The sample size}
}
\description{
Sample nsam chromosomes from population
}

