\name{stackloss}
\alias{stackloss}
\non_function{}
\title{Stackloss data}
\usage{data(stackloss)}
\description{Brownlee?s stack loss data. \cr
The \code{stackloss} data frame has 21 rows and 4 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Air}{Air flow }
    \item{Temp}{Cooling water inlet temperature}
    \item{Conc}{Acid concentration}
    \item{Loss}{Stack loss}
  }
}
\details{}
\source{}
\references{Atkinson, A.C. and Riani, M. (2000),  \emph{Robust Diagnostic Regression Analysis}, First Edition. New York: Springer, Table A.5}
\examples{}
\keyword{datasets}
