\name{wool}
\alias{wool}
\non_function{}
\title{Wool data}
\usage{data(wool)}
\description{
Number of cycles to failure of samples of worsted yarn in a 33 experiment. \cr
The \code{wool} data frame has 27 rows and 4 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{x1}{factor levels: -1, 0, 1}
    \item{x2}{factor levels: -1, 0, 1}
    \item{x3}{factor levels: -1, 0, 1}
    \item{y}{cycles to failure a numeric vector}
  }
}
\details{}
\source{}
\references{Atkinson, A.C. and Riani, M. (2000),  \emph{Robust Diagnostic Regression Analysis}, First Edition. New York: Springer,  Table A.3}
\examples{}
\keyword{datasets}
