\name{carinsuk}
\alias{carinsuk}
\docType{data}
\title{Car insurance data}
\usage{data(carinsuk)}
\description{
The \code{carinsuk} data frame has 128 rows and 5 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{OwnerAge}{a factor with levels: \code{17-20}, \code{21-24}, \code{25-29}, \code{30-34}, \code{35-39}, \code{40-49}, \code{50-59}, \code{60+}}
    \item{Model}{a factor with levels: \code{A}, \code{B}, \code{C}, \code{D} }
    \item{CarAge}{a factor with levels: \code{0-3}, \code{10+}, \code{4-7}, \code{8-9} }
    \item{NClaims}{a numeric vector}
    \item{AvCost}{a numeric vector}
  }
}
%\details{}
%\source{}
\references{Atkinson, A.C. and Riani, M. (2000),  \emph{Robust Diagnostic Regression Analysis}, First Edition. New York: Springer, Table A.16}
%\examples{}
\keyword{datasets}
