\name{carr}
\alias{carr}
\non_function{}
\title{n-Pentane data}
\usage{data(carr)}
\description{Reaction rate for Catalytic Isomerization of n-Pentane to
Isopentane \cr
The \code{carr} data frame has 24 rows and 4 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{x1}{partial pressure of hydrogen}
    \item{x2}{partial pressure of n-pentane}
    \item{x3}{partial pressure of iso-pentane}
    \item{y}{rate of disappearance of n-pentane}
  }
}
\details{}
\source{}
\references{Atkinson, A.C. and Riani, M. (2000),  \emph{Robust Diagnostic Regression Analysis}, First Edition. New York: Springer, Table A.15}
\examples{}
\keyword{datasets}
