\name{derailme}
\alias{derailme}
\non_function{}
\title{British Train Accidents.}
\usage{data(derailme)}
\description{
The \code{derailme} data frame has 67 rows and 5 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Month}{a numeric vector}
    \item{Year}{a numeric vector}
    \item{Type}{a numeric vector}
    \item{TrainKm}{a numeric vector}
    \item{y}{a numeric vector}
  }
}
\details{}
\source{}
\references{Atkinson, A.C. and Riani, M. (2000),  \emph{Robust Diagnostic Regression Analysis}, First Edition. New York: Springer, Table A.18}
\examples{}
\keyword{datasets}
