% Generated by roxygen2 (4.0.1): do not edit by hand
\name{evolveWFsingle}
\alias{evolveWFsingle}
\title{Evolve a single population under multiplicative fitness and arbitrary changes in N}
\usage{
evolveWFsingle(N0, mu_n, mu_s, s, h, littler, Ns, seed, dist = TRUE)
}
\arguments{
\item{N0}{The initial population size}

\item{mu_n}{The neutral mutation rate (per gamete per generation)}

\item{mu_n}{The mutation rate to selected mutations (per gamete per generation)}

\item{s}{The selection coefficient (for fixed-s models), or the mean of an exponential (for models with a distribution)}

\item{h}{The dominance of a selected mutation}

\item{littler}{The recombination rate (per diploid, per generation).}

\item{Ns}{A list of populaton sizes.  The length of the list is the number of generations to simulate}

\item{seed}{Random number seed for the simulation}

\item{dist}{If true, simulate a distribution of fitness effects, otherwise treat s as fixed}
}
\description{
Evolve a single population under multiplicative fitness and arbitrary changes in N
}

