\name{lakes}
\alias{lakes}
\non_function{}
\title{Lakes data}
\usage{data(lakes)}
\description{
The \code{lakes} data frame has 29 rows and 3 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{x1}{average influent nitrogenon concentration}
    \item{x2}{a numeric vector}
    \item{y}{a numeric vector}
  }
}
\details{}
\source{}
\references{Atkinson, A.C. and Riani, M. (2000),  \emph{Robust Diagnostic Regression Analysis}, First Edition. New York: Springer, Table A.14}
\examples{}
\keyword{datasets}
