\name{salinity}
\alias{salinity}
\docType{data}
\title{Salinity data}
\usage{data(salinity)}
\description{
The \code{salinity} data frame has 28 rows and 4 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{lagsalinity}{Lagged salinity}
    \item{trend}{Trend}
    \item{waterflow}{Water flow}
    \item{salinity}{Salinity}
  }
}
%\details{}
%\source{}
\references{Atkinson, A.C. and Riani, M. (2000),  \emph{Robust Diagnostic Regression Analysis}, First Edition. New York: Springer, Table A.6}
%\examples{}
\keyword{datasets}
