% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sampleStd}
\alias{sampleStd}
\title{Sample nsam chromosomes from population}
\usage{
sampleStd(pop, rng, nsam, remove_fixed = FALSE)
}
\arguments{
\item{pop}{A population returned from evolve_std}

\item{rng}{A GSLrng}

\item{nsam}{The sample size}

\item{remove_fixed}{If TRUE, do not include mutations fixed in the sample in the output.  Default is to include fixations}
}
\description{
Sample nsam chromosomes from population
}

